Avoidance Movement -- Inferno Key Setting
Added Inferno Key Setting

Changed Avoid/Kite Movement so that it will use the location for a
specific amount of time. This means it will continue to move until
either it is blocked or it reaches the location. (This fixes the jerky
movement since it would only use the location for one loop.)