Cluster Info
ClusterInfo is a class that tracks data of units that are passed thru
during the clustering process.

Cluster conditions is an enum of flags that is used to describe general
things about the cluster.
--Weak/Strong, Large/Small, Elites/Boss, Fast/Ranged

Grouping now uses this cluster info in triggering the look for groups,
and the groups that are valid to engage.