Line of Sight Update
Added melee code for LOS check to single target conditions

Added new Class LOSConditions -- used to track what tests we should be
conducting based on the current abilities being used.

LOSConditions added to Player class -- and method to construct which is
called at the end of recreation.

Added check for cluster exception units to verify that its not
clustered. (so it will use only single target conditions on exceptions
that are not clustered)

Misc:
Removed last usage of MainGridProvider.Update()